+++ packages/cli/core/src/config/types/index.ts
@@ -13,6 +13,7 @@ import type {
   BasePluginOptions,
   TerserOptions as RawTerserOptions,
 } from 'terser-webpack-plugin';
+import type { AcceptedPlugin as PostCSSPlugin } from 'postcss';
 import type { PluginConfig } from '../../loadPlugins';
 import type { TestConfig, JestConfig } from './test';
 import type { SassConfig, SassLoaderOptions } from './sass';
@@ -207,6 +208,8 @@ export type RequestHandler = (
 ) => void;
 
 export type DevServerConfig = {
+  hot?: boolean;
+  liveReload?: boolean;
   proxy?: BffProxyOptions;
   headers?: Record<string, string>;
   before?: RequestHandler[];
@@ -216,7 +219,10 @@ export type DevServerConfig = {
 
 export type PostCSSConfig =
   | PostCSSLoaderOptions
-  | ((options: PostCSSLoaderOptions) => PostCSSLoaderOptions | void);
+  | ((
+      options: PostCSSLoaderOptions,
+      utils: { addPlugins: (plugins: PostCSSPlugin | PostCSSPlugin[]) => void },
+    ) => PostCSSLoaderOptions | void);
 
 export type WebpackConfig =
   | WebpackConfiguration
@@ -313,12 +319,12 @@ export interface RuntimeByEntriesConfig {
   [name: string]: RuntimeConfig;
 }
 
-export type BffConfig = Partial<{
-  prefix: string;
-  requestCreator: string;
-  fetcher: string;
-  proxy: Record<string, any>;
-}>;
+export interface BffConfig {
+  prefix?: string;
+  requestCreator?: string;
+  fetcher?: string;
+  proxy?: Record<string, any>;
+}
 
 export interface UserConfig {
   source?: SourceConfig;