+++ packages/cli/plugin-jarvis/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-jarvis
 
+## 1.2.13
+
+### Patch Changes
+
+- a1198d509: feat: bump babel 7.18.0
+- Updated dependencies [a1198d509]
+  - @modern-js-app/eslint-config@1.2.13
+
 ## 1.2.11
 
 ### Patch Changes