+++ packages/server/adapter-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/adapter-helpers
 
+## 1.2.6
+
+### Patch Changes
+
+- a1198d509: feat: bump babel 7.18.0
+
 ## 1.2.5
 
 ### Patch Changes