+++ packages/toolkit/compiler/style/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/style-compiler
 
+## 1.2.9
+
+### Patch Changes
+
+- a1198d509: feat: bump babel 7.18.0
+
+## 1.2.8
+
+### Patch Changes
+
+- eedece8b8: relativeUrls set true
+
 ## 1.2.7
 
 ### Patch Changes